requirements.txt
# Core
asyncpg>=0.29.0,<0.30
anthropic>=0.30.0,<0.40
pydantic>=2.6.0,<3.0
langgraph>=0.1.0,<0.3

# Data generation
faker>=24.0.0,<26.0
reportlab>=4.2.0,<5.0
openpyxl>=3.1.0,<4.0

# MCP server (Phase 5)
fastmcp>=0.9.0
fastapi>=0.115.0,<1.0
uvicorn>=0.30.0,<1.0
httpx>=0.27.0,<0.29
redis>=5.0.0,<6.0
PyJWT[crypto]>=2.8.0,<3.0

# Dev / test
python-dotenv>=1.0.0,<2.0
pytest>=8.0.0,<9.0
pytest-asyncio>=0.23.0,<0.25
fakeredis>=2.23.0,<3.0

# Your Week 3 pipeline (uncomment the one you used):
# mineru>=1.0
# docling>=2.0
